[protocol] Move fixidity.ts to utils
So celotool can use it too.
