+++ packages/celotool/package.json
@@ -6,7 +6,6 @@
   "author": "Celo",
   "license": "Apache-2.0",
   "dependencies": {
-    "@celo/protocol": "^1.0.0",
     "@celo/verification-pool-api": "^1.0.0",
     "@celo/utils": "^0.0.6-beta5",
     "@celo/walletkit": "^0.0.14",
