+++ packages/protocol/package.json
@@ -14,7 +14,6 @@
     "test": "node runTests.js",
     "test:coverage": "yarn run test --coverage",
     "gas": "yarn run test:local --gas",
-    "postinstall": "yarn run build && rm -f node_modules/web3/index.d.ts",
     "build:ts": "rm -f migrations/*.js* && tsc -b",
     "build": "ts-node ./scripts/build.ts && yarn build:ts",
     "migrate": "./scripts/bash/migrate.sh",
